requirements.txt
ipykernel
langchain
langchain-community
pypdf
python-dotenv
bs4
wikipedia
arxiv
langchain-text-splitters
langchain-openai
chromadb
langchain_huggingface
faiss-cpu
langchain_chroma
langchain_groq
langchain_core
langserve
